feat: move cloud provider regions method to cluster
