+++ mongodbatlas/mongodbatlas.go
@@ -136,7 +136,6 @@ type Client struct {
 	PerformanceAdvisor                  PerformanceAdvisorService
 	CloudProviderAccess                 CloudProviderAccessService
 	DefaultMongoDBMajorVersion          DefaultMongoDBMajorVersionService
-	CloudProviderRegions                CloudProviderRegionsService
 	IPInfo                              IPInfoService
 
 	onRequestCompleted RequestCompletionCallback
@@ -288,7 +287,6 @@ func NewClient(httpClient *http.Client) *Client {
 	c.PerformanceAdvisor = &PerformanceAdvisorServiceOp{Client: c}
 	c.CloudProviderAccess = &CloudProviderAccessServiceOp{Client: c}
 	c.DefaultMongoDBMajorVersion = &DefaultMongoDBMajorVersionServiceOp{Client: c}
-	c.CloudProviderRegions = &CloudProviderRegionsServiceOp{Client: c}
 	c.IPInfo = &IPInfoServiceOp{Client: c}
 
 	return c
